implement Caveat for resset.Action
this allows restricting access without specifying a resource
