+++ caveat.go
@@ -39,6 +39,7 @@ const (
 	AttestationAuthFlyioUserID
 	AttestationAuthGitHubUserID
 	AttestationAuthGoogleUserID
+	CavAction
 
 	// allocate internal blocks of size 255 here
 	block255Min    CaveatType = 1 << 16
